+++ core/src/main/java/react4j/annotations/ReactComponent.java
@@ -54,17 +54,6 @@ public @interface ReactComponent
   String name() default "<default>";
 
   /**
-   * Indicate whether the generated component class is expected to be created and injected by the injection framework.
-   * {@link Feature#ENABLE} will force injection framework integration, {@link Feature#DISABLE}
-   * will force no injection framework integration and {@link Feature#AUTODETECT} will enable injection framework
-   * integration  if any fields or methods in the component or any parent type has an javax.inject.Inject
-   * annotation.
-   *
-   * @return enum controlling injection framework integration.
-   */
-  Feature inject() default Feature.AUTODETECT;
-
-  /**
    * Enum indicating the capabilities of the component.
    * See the {@link Type} enum for further details.
    *